refactor: move sparse and apply function to utils
